mysite03 - guestbook-spa # scroll 작업중

+++ mysite03/src/main/java/com/poscoict/mysite/service/GuestbookService.java
@@ -15,7 +15,15 @@ public class GuestbookService {
 		return guestbookRepository.findAll();
 	}
 	
+	public List<GuestbookVo> getMessageList(Long no) {
+		return guestbookRepository.findByNo(no);
+	}
+	
 	public Boolean deleteMessage(Long no, String password) {
+		if (password == null || password.isBlank()) {
+			return false;
+		}
+		
 		return guestbookRepository.delete(no, password);
 	}
